+++ service/src/main/java/com/senla/cources/service/UserWallServiceImpl.java
@@ -1,5 +1,6 @@
 package com.senla.cources.service;
 
+import com.senla.cources.domain.User;
 import com.senla.cources.domain.UserWall;
 import com.senla.cources.domain.UserWallMessage;
 import com.senla.cources.dto.UserWallMessageDto;
@@ -10,10 +11,10 @@ import com.senla.cources.repository.UserRepository;
 import com.senla.cources.repository.UserWallMessageRepository;
 import com.senla.cources.repository.UserWallRepository;
 import com.senla.cources.service.interfaces.UserWallService;
+import com.senla.cources.service.security.PermissionService;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,15 +31,17 @@ public class UserWallServiceImpl implements UserWallService {
     private UserWallMessageRepository userWallMessageRepository;
     @Autowired
     private UserWallMessageMapper userWallMessageMapper;
+    @Autowired
+    private PermissionService permissionService;
 
     @Override
     @Transactional
     public void postMessage(int userId, UserWallMessageDto messageDto) {
         log.debug(String.format("Call method postMessage with userId %d", userId));
         try {
-            UserWall userWall = userRepository.findById(userId).get().getUserWall();
-            String userName = SecurityContextHolder.getContext().getAuthentication().getName();
-            if (userWall.getUser().getUserPrincipal().getUsername().equals(userName)) {
+            User user = userRepository.findById(userId).get();
+            UserWall userWall = user.getUserWall();
+            if (permissionService.checkPermissionForUserWall(user)) {
                 UserWallMessage message = userWallMessageMapper.userWallMessageDtoToUserWallMessage(messageDto);
                 userWall.getMessages().add(message);
                 if (message.getUserWall() == null) {
@@ -47,9 +50,9 @@ public class UserWallServiceImpl implements UserWallService {
                 message.setId(null);
                 userWall.getMessages().add(message);
                 userWallRepository.save(userWall);
-                log.info(String.format("Message %d successfully posted on the wall of user %s", message.getId(), userWall.getUser().getUserPrincipal().getUsername()));
+                log.info(String.format("Message %d successfully posted on the wall of user %s", message.getId(), user.getUserPrincipal().getUsername()));
             } else {
-                log.warn(String.format("User %s isn't owner of wall %d", userWall.getUser().getUserPrincipal().getUsername(), userWall.getId()));
+                log.warn(String.format("User %s doesn't have permission to post message at wall %d", user.getUserPrincipal().getUsername(), userWall.getId()));
             }
         } catch (Exception e) {
             log.error(e.getMessage());
@@ -62,13 +65,21 @@ public class UserWallServiceImpl implements UserWallService {
     public void removeMessage(int userId, int messageId) {
         log.debug(String.format("Call method removeMessage with userId %d, messageId %d", userId, messageId));
         try {
-            UserWall userWall = userRepository.findById(userId).get().getUserWall();
-            UserWallMessage message = userWallMessageRepository.findById(messageId).get();
-            userWallMessageRepository.delete(message);
-            log.info(String.format("Message %d successfully removed from the wall of user %s", message.getId(), userWall.getUser().getUserPrincipal().getUsername()));
+            User user = userRepository.findById(userId).get();
+            UserWall userWall = user.getUserWall();
+            if (permissionService.checkPermissionForUserWall(user)) {
+                UserWallMessage message = userWallMessageRepository.findById(messageId).get();
+                userWallMessageRepository.delete(message);
+                log.info(String.format("Message %d successfully removed from the wall of user %s", message.getId(), user.getUserPrincipal().getUsername()));
+            } else {
+                String message = String.format("User %s doesn't have permission to remove message at wall %d", user.getUserPrincipal().getUsername(), userWall.getId());
+                log.warn(message);
+                throw new RuntimeException(message);
+            }
         } catch (Exception e) {
             log.error(e.getMessage());
             throw new RemoveMessageException(e);
         }
     }
+
 }
